backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
boto3==1.34.162
azure-identity==1.17.1
azure-mgmt-costmanagement==7.0.0
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
prometheus-client==0.20.0
pydantic==2.9.1
APScheduler==3.10.4
pandas==2.2.2
requests==2.32.3
